requirements.txt
absl-py==1.2.0
apt-xapian-index==0.49
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asn1crypto==1.4.0
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
autocorrect==2.6.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.4
Brlapi==0.8.3
cachetools==5.2.0
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==4.1.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==2.3.0
django-pandas==0.6.6
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
duplicity==0.8.21
entrypoints==0.4
executing==1.0.0
fasteners==0.14.1
fastjsonschema==2.16.1
filelock==3.6.0
flatbuffers==1.12
fuse-python==1.0.2
future==0.18.2
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
gTTS==2.2.4
h5py==3.7.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==1.7.0
inflection==0.5.1
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itypes==1.2.0
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.15.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==2.9.0
Keras-Preprocessing==1.1.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==14.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mechanize==0.4.7
mistune==2.0.4
monotonic==1.6
more-itertools==8.10.0
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
netifaces==0.11.0
nltk==3.7
notebook==6.4.12
numpy==1.23.2
oauthlib==3.2.0
olefile==0.46
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.12.4
psutil==5.9.1
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.20
pycups==2.0.1
Pygments==2.13.0
PyGObject==3.42.1
PyJWT==2.3.0
pylibacl==0.6.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
pytz==2020.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
regex==2022.8.17
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.2
scipy==1.9.1
SecretStorage==3.3.1
Send2Trash==1.8.0
six==1.15.0
sklearn==0.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
sqlparse==0.3.1
stack-data==0.5.0
systemd-python==234
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
whitenoise==6.2.0
widgetsnbextension==4.0.3
wrapt==1.14.1
xdg==5
xkit==0.0.0
zipp==3.1.0
